old-tests/submission/Microsoft/semantic/footer.html

<!DOCTYPE html>
<html><head>
  <meta content="text/html; charset=utf-8" http-equiv="content-type">
  <title>verify FOOTER is a HTMLElement</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 <link href="http://www.w3.org/TR/html5/the-footer-element.html#the-footer-element" rel="help">
</head>
<body>
<div id="log"></div>
 <article>
  <footer id="footer">
   <p>P in an article and footer</p>
  </footer>
 </article>
<script type="text/javascript">
 test(function()
 {
    assert_array_equals( "[object HTMLElement]",
                         document.getElementById("footer").toString() )
 })
</script>


</body></html>
